Added a dependency on pyyaml
Moved the parsing to learn.py and modified it a bit

Split some code into files, murdered most of utils

Changed config management to .yaml

Changed config behavior in the environment

Added obs/action space in the environment
