+++ configs/open_map.yaml
@@ -1,0 +1,24 @@
+---
+env_config:
+  num_agents: 10
+  size: !!python/tuple [22, 42]
+  sight_width: 5
+  sight_dist: 10
+  num_crosses: 50
+  apple_values_method:
+  tagging_values_method:
+  sustainability_metric: deepmind
+ray_config:
+  framework: torch
+  train_batch_size: 4000
+  sgd_minibatch_size: 128
+  num_sgd_iter: 3
+  lambda: 0.92
+  kl_coeff: 0.2
+  lr: 1.0e-5
+  vf_loss_coeff: 1.0e-2
+  gamma: 0.99
+  callbacks: CPRCallbacks
+run_config:
+  heterogeneous: true
+  verbose: true
