+++ environment.yaml
@@ -26,3 +26,4 @@ dependencies:
         - numpy==1.20.1
         - gputil
         - typarse
+        - pyyaml
