+++ requirements.txt
@@ -8,4 +8,9 @@ torch==1.8.1
 ray[rllib]==1.2.0
 wandb
 gputil
-typarse
+typarse~=3.3.0
+pytest~=6.2.2
+hypothesis~=6.8.1
+pyyaml~=5.4.1
+configparser~=5.0.2
+gym~=0.18.0
